Move WalletProcedure to a separate file
