+++ packages/wallet/src/redux/channelState/challenging/reducer.ts
@@ -1,6 +1,6 @@
 import * as states from '../state';
 import * as actions from '../../actions';
-import { WalletAction, WalletProcedure } from '../../actions';
+import { WalletAction } from '../../actions';
 import { unreachable } from '../../../utils/reducer-utils';
 import { createForceMoveTransaction } from '../../../utils/transaction-generator';
 import {
@@ -11,6 +11,7 @@ import {
 import { handleSignatureAndValidationMessages } from '../../../utils/state-utils';
 import { bigNumberify } from 'ethers/utils';
 import { StateWithSideEffects } from '../../utils';
+import { WalletProcedure } from '../../types';
 
 export const challengingReducer = (
   state: states.ChallengingState,
