+++ packages/wallet/src/redux/fundingState/directFunding/reducer.ts
@@ -8,7 +8,7 @@ import { StateWithSideEffects } from 'src/redux/utils';
 import { depositingReducer } from './depositing/reducer';
 import { bigNumberify } from 'ethers/utils';
 import { createDepositTransaction } from '../../../utils/transaction-generator';
-import { WalletProcedure } from '../../actions';
+import { WalletProcedure } from '../../types';
 
 export const directFundingStatusReducer = (
   state: states.DirectFundingStatus,
