+++ packages/wallet/src/redux/types.ts
@@ -1,0 +1,4 @@
+export const enum WalletProcedure {
+  DirectFunding = 'DirectFunding',
+  IndirectFunding = 'IndirectFunding',
+}
